requirements.txt
autopep8==1.5.5
djitellopy2==2.2
numpy==1.20.1
opencv-python==4.5.1.48
pycodestyle==2.6.0
pygame==2.0.1
toml==0.10.2
